Phase 6 : gérer les catégories.

+++ src/Repository/CategorieRepository.php
@@ -23,6 +23,12 @@ class CategorieRepository extends ServiceEntityRepository
 
     public function add(Categorie $entity, bool $flush = false): void
     {
+        foreach ($this->findAll() as $categorie) {
+            $name = $categorie->getName();
+            if($entity->getName()== $name){
+                return;
+            }
+        }
         $this->getEntityManager()->persist($entity);
 
         if ($flush) {
@@ -32,7 +38,9 @@ class CategorieRepository extends ServiceEntityRepository
 
     public function remove(Categorie $entity, bool $flush = false): void
     {
-        $this->getEntityManager()->remove($entity);
+        if($entity->getFormations()->isEmpty()){
+                $this->getEntityManager()->remove($entity);
+            }
 
         if ($flush) {
             $this->getEntityManager()->flush();
@@ -53,6 +61,17 @@ class CategorieRepository extends ServiceEntityRepository
                 ->orderBy('c.name', 'ASC')   
                 ->getQuery()
                 ->getResult();        
-    }    
+    }
+    
+    /**
+     * Retourne toutes des catégories 
+     * @return array
+     */
+    public function findAll(): array{
+        return $this->createQueryBuilder('c') 
+                ->orderBy('c.name', 'ASC')   
+                ->getQuery()
+                ->getResult();        
+    }
 
 }
